Refactor utilities in separate files
