+++ content-scripts/src/modules/options/timeline.js
@@ -1,4 +1,5 @@
-import { addStyles, removeElement } from "../utilities";
+import addStyles from "../utilities/addStyles";
+import removeElement from "../utilities/removeElement";
 
 // Function to change Timeline Width
 export const changeTimelineWidth = (timelineWidth) => {
