+++ content-scripts/src/modules/utilities/addStyles.js
@@ -1,0 +1,11 @@
+// Utility function to inject CSS into page
+export default function addStyles(id, css) {
+  // First remove before adding
+  removeElement(id);
+
+  const head = document.querySelector("head");
+  const style = document.createElement("style");
+  style.id = id;
+  style.textContent = `${css}`;
+  head.appendChild(style);
+}
