+++ content-scripts/src/modules/utilities/constructNewData.js
@@ -1,0 +1,21 @@
+// Utility function to create data for `injectAllChanges()`
+export default function constructNewData(changes) {
+  // Creates an array of objects from changes
+  // The value of each object is the new value
+  const newValuesArray = Object.entries(changes).map((item) => {
+    const itemKey = item[0];
+    const itemValue = item[1]?.newValue;
+    return { [itemKey]: itemValue };
+  });
+
+  // Recreate a hash map to pass to `injectAllChanges()`
+  const newChangesData = Object.fromEntries(
+    newValuesArray.map((item) => {
+      const itemKey = Object.keys(item)[0];
+      const itemValue = Object.values(item)[0];
+      return [itemKey, itemValue];
+    })
+  );
+
+  return newChangesData;
+}
