+++ content-scripts/src/modules/utilities/removeElement.js
@@ -1,0 +1,5 @@
+// Utility function to remove DOM element
+export default function removeElement(id) {
+  const element = document.getElementById(id);
+  element && element.remove();
+}
