+++ extension/content/main.css
@@ -160,7 +160,7 @@ a[href*="quick_promote_web"] {
 }
 
 #mt-listsButtonNode:hover > div {
-  background-color: rgba(107, 114, 128, 0.2);
+  background-color: rgba(var(--glyphs-color-rgb), 0.1);
 }
 
 /* Don't show label at smaller screens */
