+++ extension/content/typefully.css
@@ -50,7 +50,7 @@
   font-weight: 400 !important;
 }
 #mt-typefullyGrowButton:hover > div {
-  background-color: rgba(107, 114, 128, 0.2);
+  background-color: rgba(var(--glyphs-color-rgb), 0.1);
 }
 @media only screen and (max-width: 1264px) {
   #mt-typefullyGrowButton > div > div[dir] {
